fix: Make the time error stay furing rerender
