+++ components/shared/GetStarted/fields/time-fields/DayPeriod.tsx
@@ -20,17 +20,23 @@ import { Popover, PopoverContent, PopoverTrigger } from "./TimePopover";
 import { cn } from "@/lib/utils";
 import { Check, ChevronsUpDown } from "lucide-react";
 import { getStartedForm } from "@/data/get-started-data";
-import { HourFieldProps } from "@/types/get-started";
+import { DayPeriodFieldProps } from "@/types/get-started";
 
-export function HourField({ control, type, form, day }: HourFieldProps) {
+export function DayPeriodField({
+  control,
+  type,
+  form,
+  day,
+  timeError
+}: DayPeriodFieldProps) {
   return (
     <FormField
       control={control}
       name={`availability.${day.value}${
-        type === "start" ? ".startTime.hour" : ".endTime.hour"
-      }`} // sets the name to the specific form data value
+        type === "start" ? ".startTime.dayPeriod" : ".endTime.dayPeriod"
+      }`}
       render={({ field }) => (
-        <FormItem className="flex flex-col">
+        <FormItem className="flex flex-col ml-[3px]">
           <Popover>
             <PopoverTrigger asChild>
               <FormControl>
@@ -50,29 +56,30 @@ export function HourField({ control, type, form, day }: HourFieldProps) {
             <PopoverContent className="w-16 p-0">
               <Command>
                 <CommandList>
-                  <CommandEmpty>No hours found.</CommandEmpty>
+                  <CommandEmpty>No dayPeriod found.</CommandEmpty>
                   <CommandGroup>
-                    {getStartedForm.hours.map((hour) => (
+                    {getStartedForm.dayPeriods.map((dayPeriod) => (
                       <CommandItem
-                        value={hour.label}
-                        key={hour.value}
+                        value={dayPeriod.label}
+                        key={dayPeriod.value}
                         onSelect={() => {
                           form.setValue(
                             `availability.${day.value}${
                               type === "start"
-                                ? ".startTime.hour"
-                                : ".endTime.hour"
+                                ? ".startTime.dayPeriod"
+                                : ".endTime.dayPeriod"
                             }`,
-                            hour.value
-                          ); // sets the hour value selected on the popout on the form
+                            dayPeriod.value
+                          ); // sets the name to the specific form data value
+                          timeError()
                         }}
                         className=" cursor-pointer"
                       >
-                        {hour.label}
+                        {dayPeriod.label}
                         <Check
                           className={cn(
                             "ml-[-5px]",
-                            hour.value === field.value
+                            dayPeriod.value === field.value
                               ? "opacity-100"
                               : "opacity-0"
                           )}
