+++ components/shared/GetStarted/fields/time-fields/MinuteField.tsx
@@ -20,22 +20,18 @@ import { Popover, PopoverContent, PopoverTrigger } from "./TimePopover";
 import { cn } from "@/lib/utils";
 import { Check, ChevronsUpDown } from "lucide-react";
 import { getStartedForm } from "@/data/get-started-data";
-import { DayPeriodFieldProps } from "@/types/get-started";
+import { getStartedMentorFormValues } from "@/types/get-started";
+import { MinuteFieldProps } from "@/types/get-started";
 
-export function DayPeriodField({
-  control,
-  type,
-  form,
-  day,
-}: DayPeriodFieldProps) {
+export function MinuteField({ control, type, form, day, timeError }: MinuteFieldProps) {
   return (
     <FormField
       control={control}
       name={`availability.${day.value}${
-        type === "start" ? ".startTime.dayPeriod" : ".endTime.dayPeriod"
-      }`}
+        type === "start" ? ".startTime.minute" : ".endTime.minute"
+      }`} // sets the name to the specific form data value
       render={({ field }) => (
-        <FormItem className="flex flex-col ml-[3px]">
+        <FormItem className="flex flex-col">
           <Popover>
             <PopoverTrigger asChild>
               <FormControl>
@@ -55,29 +51,30 @@ export function DayPeriodField({
             <PopoverContent className="w-16 p-0">
               <Command>
                 <CommandList>
-                  <CommandEmpty>No dayPeriod found.</CommandEmpty>
+                  <CommandEmpty>No minutes found.</CommandEmpty>
                   <CommandGroup>
-                    {getStartedForm.dayPeriods.map((dayPeriod) => (
+                    {getStartedForm.minutes.map((minute) => (
                       <CommandItem
-                        value={dayPeriod.label}
-                        key={dayPeriod.value}
+                        value={minute.label}
+                        key={minute.value}
                         onSelect={() => {
                           form.setValue(
                             `availability.${day.value}${
                               type === "start"
-                                ? ".startTime.dayPeriod"
-                                : ".endTime.dayPeriod"
+                                ? ".startTime.minute"
+                                : ".endTime.minute"
                             }`,
-                            dayPeriod.value
-                          ); // sets the name to the specific form data value
+                            minute.value
+                          ); // sets the minute value selected on the popout on the form
+                          timeError()
                         }}
                         className=" cursor-pointer"
                       >
-                        {dayPeriod.label}
+                        {minute.label}
                         <Check
                           className={cn(
                             "ml-[-5px]",
-                            dayPeriod.value === field.value
+                            minute.value === field.value
                               ? "opacity-100"
                               : "opacity-0"
                           )}
