+++ components/shared/GetStarted/fields/time-fields/Time.tsx
@@ -5,16 +5,20 @@ import { HourField } from "./hourField";
 import { MinuteField } from "./MinuteField";
 import { DayPeriodField } from "./DayPeriod";
 import { parseTime } from "@/lib/utils";
+import { useState } from "react";
 
 export function Time({ control, errors, form, day }: TimeProp) {
   const dayAvailability = form.watch(`availability.${day.value}`);
+
+  const [isError, setIsError] = useState(false);
 
   const timeError = () => {
     if (!dayAvailability) return true; // Skip days with no availability
     const startTime = parseTime(dayAvailability.startTime);
     const endTime = parseTime(dayAvailability.endTime);
-    return startTime >= endTime;
+    setIsError(startTime >= endTime);
   };
+
   return (
     <div className="flex flex-col gap-0.5">
       <div className="font-medium font-sm">{day.label}</div>
@@ -22,37 +26,63 @@ export function Time({ control, errors, form, day }: TimeProp) {
         <div className="flex gap-2 items-center">
           <div className="font-medium text-zinc-500 text-sm">Start Time</div>
           <div className="flex gap-2 items-center">
-            <HourField control={control} type="start" form={form} day={day} />
+            <HourField
+              control={control}
+              type="start"
+              form={form}
+              day={day}
+              timeError={timeError}
+            />
             <div className="text-2xl text-neutral-500 mb-1 ml-[-4px] mr-[-5px]">
               :
             </div>
-            <MinuteField control={control} type="start" form={form} day={day} />
+            <MinuteField
+              control={control}
+              type="start"
+              form={form}
+              day={day}
+              timeError={timeError}
+            />
             <DayPeriodField
               control={control}
               type="start"
               form={form}
               day={day}
+              timeError={timeError}
             />
           </div>
         </div>
         <div className="flex gap-2 items-center">
           <div className="font-medium text-zinc-500 text-sm">End Time</div>
           <div className="flex gap-2 items-center">
-            <HourField control={control} type="end" form={form} day={day} />
+            <HourField
+              control={control}
+              type="end"
+              form={form}
+              day={day}
+              timeError={timeError}
+            />
             <div className="text-2xl text-neutral-500 mb-1 ml-[-4px] mr-[-5px]">
               :
             </div>
-            <MinuteField control={control} type="end" form={form} day={day} />
+            <MinuteField
+              control={control}
+              type="end"
+              form={form}
+              day={day}
+              timeError={timeError}
+            />
             <DayPeriodField
               control={control}
               type="end"
               form={form}
               day={day}
+              timeError={timeError}
             />
           </div>
         </div>
       </div>
-      {timeError() && (
+      {isError && (
         <p className="text-red-500">Start Time can't be less than End Time</p>
       )}
     </div>
