+++ components/shared/GetStarted/fields/time-fields/hourField.tsx
@@ -20,15 +20,14 @@ import { Popover, PopoverContent, PopoverTrigger } from "./TimePopover";
 import { cn } from "@/lib/utils";
 import { Check, ChevronsUpDown } from "lucide-react";
 import { getStartedForm } from "@/data/get-started-data";
-import { getStartedMentorFormValues } from "@/types/get-started";
-import { MinuteFieldProps } from "@/types/get-started";
+import { HourFieldProps } from "@/types/get-started";
 
-export function MinuteField({ control, type, form, day }: MinuteFieldProps) {
+export function HourField({ control, type, form, day, timeError }: HourFieldProps) {
   return (
     <FormField
       control={control}
       name={`availability.${day.value}${
-        type === "start" ? ".startTime.minute" : ".endTime.minute"
+        type === "start" ? ".startTime.hour" : ".endTime.hour"
       }`} // sets the name to the specific form data value
       render={({ field }) => (
         <FormItem className="flex flex-col">
@@ -51,29 +50,30 @@ export function MinuteField({ control, type, form, day }: MinuteFieldProps) {
             <PopoverContent className="w-16 p-0">
               <Command>
                 <CommandList>
-                  <CommandEmpty>No minutes found.</CommandEmpty>
+                  <CommandEmpty>No hours found.</CommandEmpty>
                   <CommandGroup>
-                    {getStartedForm.minutes.map((minute) => (
+                    {getStartedForm.hours.map((hour) => (
                       <CommandItem
-                        value={minute.label}
-                        key={minute.value}
+                        value={hour.label}
+                        key={hour.value}
                         onSelect={() => {
                           form.setValue(
                             `availability.${day.value}${
                               type === "start"
-                                ? ".startTime.minute"
-                                : ".endTime.minute"
+                                ? ".startTime.hour"
+                                : ".endTime.hour"
                             }`,
-                            minute.value
-                          ); // sets the minute value selected on the popout on the form
+                            hour.value
+                          ); // sets the hour value selected on the popout on the form
+                          timeError()
                         }}
                         className=" cursor-pointer"
                       >
-                        {minute.label}
+                        {hour.label}
                         <Check
                           className={cn(
                             "ml-[-5px]",
-                            minute.value === field.value
+                            hour.value === field.value
                               ? "opacity-100"
                               : "opacity-0"
                           )}
